Add the CARB OBD PDF to the main folder
This is the document that states which vehicle parameters are mandated to be able to be pulled from cars. Vehicle speed (which is needed to keep track of car mileage) is one of them. On page 107, it says:

(4.2.1) For all vehicles: calculated load value, number of stored confirmed fault codes, engine coolant temperature, engine speed, absolute throttle position (if equipped with a throttle), vehicle speed, OBD requirements to which the engine is certified (e.g., California OBD II, EPA OBD, European OBD, non-OBD) and MIL status (i.e., commanded-on or commanded-off).